expose internals of Text.Earley.Parser
see https://github.com/ollef/Earley/issues/8

tests pass, docs build.

I don't think the benchmarks are affected (wrt inlining or anything)
