+++ Earley.cabal
@@ -23,7 +23,7 @@ source-repository    head
   location: https://github.com/ollef/Earley.git
 
 library
-  exposed-modules:     Text.Earley.Derived, Text.Earley.Grammar, Text.Earley.Parser Text.Earley
+  exposed-modules:     Text.Earley.Derived, Text.Earley.Grammar, Text.Earley.Parser, Text.Earley, Text.Earley.Internal
   -- other-modules:
   build-depends:       base >=4.7 && <4.9, containers >=0.5, ListLike >=4.1
   -- hs-source-dirs:
